pyproject.toml
[tool.poetry]
name = "readgpt"
version = "0.1.0"
description = "ReadGPT - Summarize and chat with any article using AI"
authors = ["Feliciano Lima <fsfelicianosantana@gmail.com>"]
readme = "README.md"
packages = [{ include = "main.py" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
ruff = "^0.9.6"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
